+++ src/Components/Atoms/Layout/Section.tsx
@@ -6,9 +6,9 @@ import { Icon } from '../Icon';
 import { Typography } from '../Typography';
 
 export interface ISectionProps {
-  /** When set the section gets collapsable ability (optional, default true) */
-  collapsable?: boolean;
-  /** Open initially if collapsable (optional, default: true) */
+  /** When set the section gets collapsible ability (optional, default true) */
+  collapsible?: boolean;
+  /** Open initially if collapsible (optional, default: true) */
   openInitially?: boolean;
   /** Add separator at the end of the section (optional, default: true) */
   separator?: boolean;
@@ -21,26 +21,28 @@ export interface ISectionProps {
 }
 
 const Section = (props: PropsWithChildren<ISectionProps>): ReactElement => {
-  const { children, collapsable, dataTestId, level, openInitially, separator, title } = props;
+  const { children, collapsible, dataTestId, level, openInitially, separator, title } = props;
 
   const { getCollapseProps, setExpanded, isExpanded } = useCollapse({ duration: 500 });
 
   useEffect(() => {
-    setExpanded(collapsable ? openInitially === true : true);
-  }, [collapsable, openInitially]);
+    setExpanded(collapsible ? openInitially === true : true);
+  }, [collapsible, openInitially]);
 
   const handleClick = () => {
-    setExpanded((prev) => !prev);
+    if (collapsible) {
+      setExpanded((prev) => !prev);
+    }
   };
 
   return (
     <div className='gds-layout-section-container'>
       <div
-        className={classnames('section-header', { collapsable: collapsable })}
+        className={classnames('section-header', { collapsible: collapsible })}
         onClick={handleClick}
         data-testid={dataTestId}>
         <Typography.Title level={level || 2}>{title}</Typography.Title>
-        {collapsable && (
+        {collapsible && (
           <Icon
             icon={['fal', 'chevron-left']}
             size='xs'
@@ -57,7 +59,7 @@ const Section = (props: PropsWithChildren<ISectionProps>): ReactElement => {
 };
 
 Section.defaultProps = {
-  collapsable: true,
+  collapsible: true,
   openInitially: true,
   separator: true,
 };